test(backend): align integration tests with Rig architecture and unify database types
- Update extensive integration test suite to match RigClient, RigMessage, and new tokenizer patterns.
- Refactor DbStringArray and unified types to fix parity issues and NULL handling discovered during testing.
- Ensure cross-backend compatibility (Postgres/SQLite) for all test cases.
- Refactor services and database models to support unified type alignment.

+++ backend/src/services/payment/credit_service.rs
@@ -248,6 +248,7 @@ impl CreditService {
         Ok(result)
     }
 
+    #[allow(clippy::too_many_arguments)]
     pub fn add_credits(
         &self,
         conn: &mut PgConnection,
@@ -880,7 +881,7 @@ impl CreditService {
         let refund_transaction = NewCreditTransaction {
             id: DbId::new(),
             user_id,
-            amount: refund_amount as i32, // Positive for refunds
+            amount: refund_amount, // Positive for refunds
             balance_after: updated_balance.balance,
             transaction_type: "refund".to_string(),
             description_encrypted: encrypted_data.description_encrypted,
